requirements.txt
PyYAML==6.0
selenium==4.6.0
